Replace Move::from_smith_notation with Move::from_coordinate_notation.
This is essentially the same notation with the redundant parts removed.
This makes it easier to parse and more intuitive for users. Additionally
this function will be used for UCI communication.

While this is not a very common chess notation it is perfect for
computer chess since no context (i.e. the current board state) is
necessary to parse a move.